fix(object-model): arch-independent Merkle k + apply_tree_delta V4 guard
Leg-1 review fixes D and E:
- `merkle_tree_hash` computes `k` (largest power of two < n) with
  `leading_zeros` on `usize`, not a widened `u64` against `usize::BITS-1` — the
  old form underflowed on wasm32 (usize::BITS == 32). A crypto path must not
  depend on pointer width.
- `apply_tree_delta` returns `Err` on a V4 salted anchor instead of silently
  producing a V3 (salt-less) tree with a different id. Unreachable via store
  paths today, but it is public API — fail loud.

+++ crates/object-model/src/object/tree.rs
@@ -815,8 +815,10 @@ fn merkle_tree_hash(leaves: &[ContentHash]) -> ContentHash {
         1 => leaves[0],
         n => {
             // k = largest power of two strictly less than n (RFC 6962:
-            // k < n <= 2k).
-            let k = 1usize << ((usize::BITS - 1) - ((n - 1) as u64).leading_zeros());
+            // k < n <= 2k). `leading_zeros` is taken on `usize` (not a widened
+            // u64) so the shift is arch-independent — a crypto path must not
+            // depend on the pointer width (wasm32 has usize::BITS == 32).
+            let k = 1usize << ((usize::BITS - 1) - (n - 1).leading_zeros());
             let left = merkle_tree_hash(&leaves[..k]);
             let right = merkle_tree_hash(&leaves[k..]);
             let mut hasher = ContentHash::typed_hasher(TREE_V4_NODE_PREFIX, 64);
